filter- and iotaviews tests moved from P3059R2_* to main-view-test-files;
