+++ tests/std/test.lst
@@ -737,10 +737,6 @@ tests\P2693R1_text_formatting_header_thread
 tests\P2693R1_text_formatting_stacktrace
 tests\P2693R1_text_formatting_thread_id
 tests\P3016R6_inconsistent_begin_end
-tests\P3059R2_basic_istream_view_iterator_constructor
-tests\P3059R2_elements_view_iterator_sentinel_constructors
-tests\P3059R2_filter_view_iterator_sentinel_constructors
-tests\P3059R2_iota_view_iterator_sentinel_constructors
 tests\P3059R2_join_view_sentinel_constructor
 tests\P3059R2_lazy_split_view_inner_outer_iterators_constructors
 tests\P3059R2_split_view_iterator_sentinel_constructors
